Refactor to make Sonos happy
