+++ src/shared/components/palette/base-field/index.ts
@@ -5,3 +5,4 @@ export * from './field-read.component';
 export * from './field-write.component';
 export * from './field-read-label.component';
 export * from './palette-context.enum';
+export * from './payment-field.component';
